Competitor_Analysis_Agent/requirements.txt
langchain-openai
langchain-community
langchain
tavily-python
openai
python-dotenv
langchain-google-genai 